+++ .gitignore
@@ -1,3 +1,3 @@
-controller/keys.js
+src/controller/keys.js
 node_modules
 .env
