--- public/javascripts/buttonId.js
@@ -1,3 +1,0 @@
-function fnPatient(comp) {
-    let id = comp.id;
-}
