+++ src/public/images/symptom_cough.png
