+++ src/public/images/symptom_headache.png
