+++ src/public/images/symptom_high-fever.png
