+++ src/public/images/symptom_sore-troath.png
