+++ src/public/images/symptoms.png
