+++ src/public/javascripts/buttonId.js
@@ -1,0 +1,4 @@
+function fnPatient(id) {
+  //TODO
+  console.log(id);
+}
