+++ src/public/javascripts/jsNavBar.js
