+++ src/views/layouts/navBar.ejs
