+++ src/views/symptoms.ejs
